pyproject.toml
[project]
name = "loupe"
version = "0.1.0"
description = "Modular aesthetic analysis tool for anime screenshots"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.6.0",
    "torchvision",
    "open-clip-torch>=3.3.0",
    "opencv-python-headless>=4.11.0",
    "numpy>=2.1",
    "scipy>=1.14",
    "scikit-learn>=1.4",
    "scikit-image>=0.24",
    "pillow>=11.0",
    "pydantic>=2.8",
    "pydantic-settings>=2.13",
    "typer>=0.15",
    "typer-config>=1.5",
    "rich>=13.0",
    "timm>=1.0.20",
    "huggingface-hub",
    "onnxruntime-gpu>=1.21",
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "pytest-benchmark",
    "pytest-xdist",
    "pytest-cov",
    "ruff",
    "pre-commit",
    "pre-commit-uv",
    "pip-audit",
]

[project.scripts]
loupe = "loupe.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loupe"]

# --- uv: PyTorch CUDA 12.8 index routing ---

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

# --- Ruff ---

[tool.ruff]
target-version = "py313"
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TC",  # flake8-type-checking
    "RUF", # ruff-specific
    "D",   # pydocstyle
]
ignore = [
    "D100",  # missing docstring in public module (module-level already covered)
    "D104",  # missing docstring in public package (__init__.py)
    "D203",  # conflicts with D211 (no-blank-line-before-class)
    "D213",  # conflicts with D212 (multi-line-summary-second-line)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["D"]  # tests don't need docstrings enforced

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
known-first-party = ["loupe"]

# --- Pyright ---

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
include = ["src", "tests"]

[[tool.pyright.executionEnvironments]]
root = "tests"
reportPrivateUsage = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownLambdaType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnnecessaryComparison = false
reportMissingTypeStubs = false

# --- Pytest ---

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "gpu: marks tests requiring CUDA GPU",
    "slow: marks tests that are slow to run",
    "integration: marks end-to-end integration tests using real images and models",
]
